Move functions to lib.js to support new directory sync function
