+++ scripts/remote/controller/lib.js
@@ -1,7 +1,9 @@
 const { createHmac, timingSafeEqual } = require('crypto');
+const rc = require('./rc.js');
 
 const log = (...args) => console.log(new Date, ...args);
 const err = (...args) => console.error(new Date, ...args);
+const Rsync = require('rsync');
 
 const respond = (res, status, msg, mediaType = 'text/plain') => {
     res.writeHead(status, { 'Content-Type': mediaType });
@@ -20,6 +22,45 @@ const getRequestBody = (req) =>
           req.on('end', () => resolve(data));
       });
 
+
+const getSshCredentials = () => ({
+    host: rc('host', 'localhost'),
+    username: rc('user', process.env.USER),
+    privateKey: rc('private_key', `${process.env.HOME}/.ssh/id_rsa`),
+});
+
+const getRemoteHomeDirectory = () => {
+    const { username } = getSshCredentials();
+    return `/home/${username}`;
+};
+
+const syncRemoteDirectory = (srcPath, dest) => {
+    const { username, host } = getSshCredentials();
+    let src = `${username}@${host}:${srcPath}`;
+
+    // Trailing '/' is meaningful: https://serverfault.com/a/529294
+    src = src[src.length - 1] === '/' ? src : src + '/';
+
+    // TODO: Stream rsync stdin/stdout to new files.
+    log(`Syncing ${src} -> ${dest}`);
+
+    const rsync = (
+        new Rsync()
+            .flags('abziu')
+            .set('e', `ssh -v -i ${rc('private_key')}`)
+            .source(src)
+            .destination(dest)
+    );
+
+    return new Promise((resolve, reject) =>
+        rsync.execute((error, code, cmd) => {
+            if (error)
+                reject(error)
+            else
+                resolve();
+        }));
+}
+
 module.exports = {
     log,
     err,
@@ -27,4 +68,7 @@ module.exports = {
     respond,
     checkHmac,
     getRequestBody,
+    syncRemoteDirectory,
+    getRemoteHomeDirectory,
+    getSshCredentials,
 };
